// tmp_app/Api.ts
export interface output$api$videos$get_video {
    vid: string, username: string, f_name: string, l_name: string, channel: string, title: string, description: string, upload_date: string, video_path: string, image_path: string, last_access: string, category: string
}